Cluster the OSIVQ data using consensus between various algorithms

Cluster the OSIVQ data using consensus between various algorithms

Usage

cluster_osivq(
  df,
  algorithms = c("gmm", "pam", "cmeans"),
  cons.funs = c("kmodes", "majority", "CSPA"),
  seed = 667,
  progress = FALSE,
  verbose = FALSE
)

Arguments

df

A dataframe with the OSIVQ scores, typically obtained from get_clean_data().

algorithms

A character vector of clustering algorithms to use. Default is c("gmm", "pam", "cmeans", "diana"). See ?diceR::dice() for more details.

cons.funs

A character vector of consensus functions to use. Default is c("CSPA"). See ?diceR::dice() for more details.

seed

An integer seed for reproducibility. Default is 667.

progress

Logical value indicating whether to show a progress bar.

verbose

Logical value indicating whether to print detailed messages during the clustering process.

Value

A list with the clustering results from diceR::dice().